Implement timer display
We want to be able to actually see the timer counting down.
`sleep()` was insufficient for this, thus we moved to using
`clock_gettime()` and `nanosleep()`.
This should give us much more flexibility.

Currently, we simply poll the time 10 times per second, which seems
sufficient for the time being.

Display is done directly to stdout. We need to refactor soon to
facilitate displaying on multiple platforms, both for CLI and GUI.