Fix gitchy form
Upon successive use of the create timer form, visual attributes were not
being preserved.

It appears this was due to clearing the contents of the window which
houses the form.
Rather than trying to re-use the same window all the time, we have
started to create separate windows for separate purposes.

In creating a "popup" window we have also learned a little more about
how curses behaves, and have re-arranged a small part of the code to
reflect this.

TODO:
- Get rid of the global `win` WINDOW and move to creating and deleting
  windows as required
- Separate the popup window into it's own function(s) if need be