Add user input for sound file
What's a timer without the ability to make noise?!
For now we've just added a means by which to accept user input of a
filepath, to store that path and pass it onto the notification callback.

Paths are not allowed to contain any whitespace, unless the path is
contained within double quotes.

TODO:
- Escape quoted paths and remove quotes
- Basic input validation, e.g. does the file exist?
- Load and play sound