First pass on sound
Add the ability for a user to attach a sound file to a timer.
On completion, the sound is played and looped.

Unfortunately portaudio sometimes dumps to stderr and at present our
solution is simply a workaround.

libsndfile with portaudio give us a simple and nice way to play most
common audio files. They are currently required as a dependency; we may
choose to include the sources in future builds.

There is still a lot of room for better integrating audio into the app.
For one, we should have a more clean method for providing status
messages to the user.